Move transactions into btc module
