+++ sandbox_tests/src/secp256k1_hack.rs
@@ -11,7 +11,7 @@ use secp256k1::key::SecretKey;
 use secp256k1::key;
 use secp256k1::Error;
 
-use anchoring_btc_service::details::transactions::RawBitcoinTx;
+use anchoring_btc_service::details::btc::transactions::RawBitcoinTx;
 
 /// The structure with the same memory representation as the `secp256k1::Secp256k1`.
 #[derive(Clone, Copy)]
