+++ src/details/mod.rs
@@ -3,7 +3,6 @@ mod macros;
 
 pub mod btc;
 pub mod rpc;
-pub mod transactions;
 pub mod error;
 pub mod regtest;
 #[cfg(feature="sandbox_tests")]
